backend/user-requirements.txt
Flask~=2.2.2
pyrebase4==4.6.0
numpy==1.21.2
pandas~=1.3.5
scipy==1.7.2
protobuf==3.20.0
oauthlib==3.0.0
matplotlib
google-auth==1.0.0
pyyaml
absl-py==1.4.0
pytz==2017.3
networkx==3.0
flask-cors
open_gopro
asyncio==3.4.3
boxsdk==3.6.1
opencv-python==4.7.0.68
av==10.0.0
pynput==1.7.6
open3d==0.17.0
pynput
websockets
redis
pyaudio==0.2.13
pysftp==0.2.9
tqdm
mediapipe
gunicorn
urllib3==1.26.15
requests-toolbelt==0.10.1
label-studio-sdk==0.0.27
moviepy
torch~=1.13.1
torchvision~=0.14.1